Fix: UIImage Extension 파일 생성
